Checked some code
Changed MUA.exe to Game.exe. Game.exe is the correct executable to run MUA on modern Windows.
Changed an error message to better English.
Added comments about future changes
Added elevated command code for OHS (function only). Not sure if it works like that. Running it elevated directly from C# is not possible, would require the program to be run as admin. It's possible that the prompt for OHS would not appear if we ran this program as admin, but it would need to be tested first.